hashtag_rd1/prediction/requirements.txt
click
h11
pydantic
starlette
fastapi
openai
PyJWT
passlib
uvicorn
python-dotenv
pandas